requirements.txt
APScheduler==3.9.1.post1
Django==4.1.3
django-simple-history==3.2.0
mercadopago==2.2.0
num2words==0.5.12
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
weasyprint==57.1
qrcode~=7.4.1